requirements.txt
langchain>=0.2.0
langchain-core>=0.2.0
langchain-openai>=0.1.0
langfuse>=2.0.0
deepeval>=0.21.0
pytest>=8.0.0
pydantic>=2.0.0
python-dotenv>=1.0.0